7.4.1

+++ src/Module/Util/Recommendation.php
@@ -47,7 +47,7 @@ class Recommendation
      * @return SimpleXMLElement
      * @throws LastFmQueryFailedException
      */
-    public static function get_lastfm_results($method, $query): SimpleXMLElement
+    public static function get_lastfm_results(string $method, string $query): SimpleXMLElement
     {
         global $dic;
 
@@ -94,45 +94,54 @@ class Recommendation
      * @param string $type
      * @param int $object_id
      * @param bool $get_items
-     * @return array
+     * @return null|array{
+     *     id: int,
+     *     last_update: string,
+     *     items?: list<array{id: ?int, name: string, rel: ?string, mbid: ?string}>
+     * }
      */
-    protected static function get_recommendation_cache($type, $object_id, $get_items = false): array
+    protected static function get_recommendation_cache(string $type, int $object_id, bool $get_items = false): ?array
     {
         if (!AmpConfig::get('cron_cache')) {
             self::garbage_collection();
         }
 
+        $cache      = null;
         $sql        = "SELECT `id`, `last_update` FROM `recommendation` WHERE `object_type` = ? AND `object_id` = ?";
         $db_results = Dba::read($sql, [$type, $object_id]);
-
-        if ($cache = Dba::fetch_assoc($db_results)) {
-            if ($get_items) {
-                $cache['items'] = [];
-                $sql            = "SELECT `recommendation_id`, `name`, `rel`, `mbid` FROM `recommendation_item` WHERE `recommendation` = ?";
-                $db_results     = Dba::read($sql, [$cache['id']]);
-                while ($results = Dba::fetch_assoc($db_results)) {
-                    $cache['items'][] = [
-                        'id' => $results['recommendation_id'],
-                        'name' => $results['name'],
-                        'rel' => $results['rel'],
-                        'mbid' => $results['mbid'],
-                    ];
-                }
-            }
+        if ($row = Dba::fetch_assoc($db_results)) {
+            $cache = [
+                'id' => (int)$row['id'],
+                'last_update' => (string)$row['last_update'],
+            ];
         }
 
-        return $cache;
+        if (is_array($cache) && $get_items) {
+            $cache['items'] = [];
+            $sql            = "SELECT `recommendation_id` AS `id`, `name`, `rel`, `mbid` FROM `recommendation_item` WHERE `recommendation` = ?";
+            $db_results     = Dba::read($sql, [$cache['id']]);
+            while ($results = Dba::fetch_assoc($db_results)) {
+                $cache['items'][] = [
+                    'id' => ($results['id'] !== null) ? (int)$results['id'] : null,
+                    'name' => (string)$results['name'],
+                    'rel' => ($results['rel'] !== null) ? (string)$results['rel'] : null,
+                    'mbid' => ($results['mbid'] !== null) ? (string)$results['mbid'] : null,
+                ];
+            }
+
+            return $cache;
+        }
+
+        return null;
     }
 
     /**
      * delete_recommendation_cache
-     * @param string $type
-     * @param int $object_id
      */
-    protected static function delete_recommendation_cache($type, $object_id): void
+    protected static function delete_recommendation_cache(string $type, int $object_id): void
     {
         $cache = self::get_recommendation_cache($type, $object_id);
-        if (array_key_exists('id', $cache)) {
+        if ($cache !== null && array_key_exists('id', $cache)) {
             Dba::write('DELETE FROM `recommendation_item` WHERE `recommendation` = ?', [$cache['id']]);
             Dba::write('DELETE FROM `recommendation` WHERE `id` = ?', [$cache['id']]);
         }
@@ -140,11 +149,19 @@ class Recommendation
 
     /**
      * update_recommendation_cache
+     * Updates the recommendation cache for a given object type and object ID with provided recommendations.
+     *
      * @param string $type
      * @param int $object_id
-     * @param array $recommendations
+     * @param array<array{
+     *     id?: ?int,
+     *     name?: ?string,
+     *     rel?: ?string,
+     *     mbid?: ?string
+     * }> $recommendations
+     * @return void
      */
-    protected static function update_recommendation_cache($type, $object_id, $recommendations): void
+    protected static function update_recommendation_cache(string $type, int $object_id, array $recommendations): void
     {
         if (count($recommendations) > 0) {
             self::delete_recommendation_cache($type, $object_id);
@@ -170,16 +187,21 @@ class Recommendation
      * @param int $song_id
      * @param int $limit
      * @param bool $local_only
-     * @return array
+     * @return list<array{
+     *     id: ?int,
+     *     name: string,
+     *     rel: ?string,
+     *     mbid?: ?string
+     * }>
      */
-    public static function get_songs_like($song_id, $limit = 5, $local_only = true): array
+    public static function get_songs_like(int $song_id, int $limit = 5, bool $local_only = true): array
     {
         if (!AmpConfig::get('lastfm_api_key')) {
             return [];
         }
 
-        $song     = new Song($song_id);
-        $artist   = new Artist($song->artist);
+        $song   = new Song($song_id);
+        $artist = new Artist($song->artist);
 
         if ($artist->isNew()) {
             debug_event(
@@ -203,7 +225,7 @@ class Recommendation
         }
 
         $cache = self::get_recommendation_cache('song', $song_id, true);
-        if (!array_key_exists('id', $cache)) {
+        if ($cache === null || !array_key_exists('id', $cache)) {
             $similars = [];
             try {
                 $xml = self::get_lastfm_results('track.getsimilar', $query);
@@ -268,6 +290,7 @@ class Recommendation
         if ($similars) {
             $results = [];
             foreach ($similars as $similar) {
+                /** @var array{id: ?int, name: string, rel: ?string, mbid: ?string} $similar */
                 if (!$local_only || $similar['id'] !== null) {
                     $results[] = $similar;
                 }
@@ -291,16 +314,21 @@ class Recommendation
      * @param int $artist_id
      * @param int $limit
      * @param bool $local_only
-     * @return array
+     * @return list<array{
+     *     id: ?int,
+     *     name: string,
+     *     rel?: ?string,
+     *     mbid?: ?string
+     * }>
      */
-    public static function get_artists_like($artist_id, $limit = 10, $local_only = true): array
+    public static function get_artists_like(int $artist_id, int $limit = 10, bool $local_only = true): array
     {
         if (!AmpConfig::get('lastfm_api_key')) {
             return [];
         }
 
         $cache = self::get_recommendation_cache('artist', $artist_id, true);
-        if (!array_key_exists('id', $cache)) {
+        if ($cache === null || !array_key_exists('id', $cache)) {
             $artist   = new Artist($artist_id);
             $similars = [];
             $fullname = (string)$artist->get_fullname();
@@ -389,15 +417,23 @@ class Recommendation
 
     /**
      * get_artist_info_by_name
-     * Returns artist information
-     * @param string $fullname
-     * @return array
+     * Returns artist information by full name.
+     * @param string $fullname The full name of the artist.
+     * @return array{
+     *     summary: ?string,
+     *     placeformed: ?string,
+     *     yearformed: ?string
+     * }
      */
-    public static function get_artist_info_by_name($fullname): array
+    public static function get_artist_info_by_name(string $fullname): array
     {
         $query = 'artist=' . rawurlencode($fullname);
 
-        $results = [];
+        $results = [
+            'summary' => null,
+            'placeformed' => null,
+            'yearformed' => null,
+        ];
 
         try {
             $xml = self::get_lastfm_results('artist.getinfo', $query);
@@ -423,9 +459,18 @@ class Recommendation
      * get_artist_info
      * Returns artist information
      * @param int $artist_id
-     * @return array
+     * @return array{
+     *     id: ?int,
+     *     summary: ?string,
+     *     placeformed: ?string,
+     *     yearformed: ?int,
+     *     largephoto: ?string,
+     *     smallphoto: ?string,
+     *     mediumphoto: ?string,
+     *     megaphoto: ?string
+     * }
      */
-    public static function get_artist_info($artist_id): array
+    public static function get_artist_info(int $artist_id): array
     {
         $artist = new Artist($artist_id);
         $query  = ($artist->mbid)
@@ -450,10 +495,25 @@ class Recommendation
         try {
             $xml = self::get_lastfm_results('artist.getinfo', $query);
         } catch (LastFmQueryFailedException) {
-            return [];
+            return [
+                'id' => $artist_id,
+                'summary' => null,
+                'placeformed' => null,
+                'yearformed' => null,
+                'largephoto' => null,
+                'smallphoto' => null,
+                'mediumphoto' => null,
+                'megaphoto' => null,
+            ];
         }
 
-        $results            = [];
+        $results = [
+            'id' => null,
+            'largephoto' => null,
+            'smallphoto' => null,
+            'mediumphoto' => null,
+            'megaphoto' => null,
+        ];
         $results['summary'] = strip_tags(
             (string)preg_replace(
                 "#<a href=([^<]*)Last\.fm</a>.#",
@@ -487,9 +547,16 @@ class Recommendation
      * get_album_info
      * Returns album information
      * @param int $album_id
-     * @return array
+     * @return array{
+     *     id: int,
+     *     summary: ?string,
+     *     largephoto: ?string,
+     *     smallphoto: ?string,
+     *     mediumphoto: ?string,
+     *     megaphoto: ?string
+     * }
      */
-    public static function get_album_info($album_id): array
+    public static function get_album_info(int $album_id): array
     {
         $album = new Album($album_id);
         $query = ($album->mbid)
@@ -533,10 +600,8 @@ class Recommendation
 
     /**
      * Migrate an object associate stats to a new object
-     * @param string $object_type
-     * @param int $old_object_id
      */
-    public static function migrate($object_type, $old_object_id): void
+    public static function migrate(string $object_type, int $old_object_id): void
     {
         self::delete_recommendation_cache($object_type, $old_object_id);
     }
